entity framework entegre edildi.

+++ Entities/Concrete/Color.cs
@@ -7,7 +7,7 @@ namespace Entities.Concrete
 {
     public class Color:IEntity
     {
-        public int ColorId { get; set; }
+        public byte ColorId { get; set; }
         public string ColorName { get; set; }
     }
 }
